feat: implemented crypto trading use cases

+++ src/main/java/org/voxsledderman/cryptoExchange/domain/repositories/EconomyRepository.java
@@ -1,11 +1,12 @@
 package org.voxsledderman.cryptoExchange.domain.repositories;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface EconomyRepository {
 
-    double getBalance(UUID uuid);
-    boolean deposit(UUID uuid, double amount);
-    boolean withdraw(UUID uuid, double amount);
-    boolean hasEnough(UUID uuid, double amount);
+    BigDecimal getBalance(UUID uuid);
+    boolean deposit(UUID uuid, BigDecimal amount);
+    boolean withdraw(UUID uuid, BigDecimal amount);
+    boolean hasEnough(UUID uuid, BigDecimal amount);
 }
